refactor(react): update all paths to use new locations
