+++ packages/react/src/components/avatar/avatar.tsx
@@ -1,7 +1,7 @@
 import type { AvatarVariantProps } from 'styled-system/recipes'
 import type { ReactNode } from 'react'
 import { splitProps } from '../../utils'
-import { Show } from '../Show'
+import { Show } from '../show/index'
 import { AvatarParts } from './parts'
 import type { AvatarRootProps } from './primitives'
