+++ packages/react/src/components/combobox/combobox.tsx
@@ -5,8 +5,8 @@ import type { ComboboxVariantProps } from 'styled-system/recipes'
 import { useCerberusContext } from '../../context/cerberus'
 import type { SelectCollectionItem } from '../select/select'
 import { type FieldInputElProps } from '../field/index'
-import { Portal } from '../Portal'
-import { Show } from '../Show'
+import { Portal } from '../portal/index'
+import { Show } from '../show/index'
 import { ComboboxParts } from './parts'
 import { ComboboxStartIcon } from './primitives'
