+++ packages/react/src/components/tag/tag.tsx
@@ -6,7 +6,7 @@ import { css, cx } from 'styled-system/css'
 import { tag, type TagVariantProps } from 'styled-system/recipes'
 import { useCerberusContext } from '../../context/cerberus'
 import { IconButton } from '../icon-button/index'
-import { Show } from '../show/show'
+import { Show } from '../show/index'
 
 /**
  * This module contains the tag component.
