+++ packages/react/src/context/confirm-modal.tsx
@@ -14,7 +14,7 @@ import {
 import { css } from 'styled-system/css'
 import { HStack, VStack } from 'styled-system/jsx'
 import { Button } from '../components/button/button'
-import { Show } from '../components/Show'
+import { Show } from '../components/show/index'
 import { Avatar } from '../components/avatar/avatar'
 import {
   Dialog,
